% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bias_dm3}
\alias{bias_dm3}
\title{Genomic biases for dm3 (flybase)}
\format{A data frame with 130434 rows and 4 variables:
\describe{
  \item{chr}{chromosome}
  \item{position}{start postion of the genomic bin}
  \item{map}{mappability as computed by gem}
  \item{cg}{cg content as computed by bedtools}
}}
\usage{
bias_dm3
}
\description{
A \code{data.frame} containing the mappability and CG
proportion of the mouse genome (dm3, flybase) in 1 Kbp bins
}
\keyword{datasets}
