% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bias_hg38}
\alias{bias_hg38}
\title{Genomic biases for hg38}
\format{A data frame with 3031043 rows and 4 variables:
\describe{
  \item{chr}{chromosome}
  \item{position}{start postion of the genomic bin}
  \item{map}{mappability as computed by gem}
  \item{cg}{cg content as computed by bedtools}
}}
\usage{
bias_hg38
}
\description{
A \code{data.frame} containing the mappability and CG proportion
of the human genome (hg38) in 1 Kbp bins
}
\keyword{datasets}

