% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bias_mm10}
\alias{bias_mm10}
\title{Genomic biases for mm10}
\format{A data frame with 2617314 rows and 4 variables:
\describe{
  \item{chr}{chromosome}
  \item{position}{start postion of the genomic bin}
  \item{map}{mappability as computed by gem}
  \item{cg}{cg content as computed by bedtools}
}}
\usage{
bias_mm10
}
\description{
A \code{data.frame} containing the mappability and CG
proportion of the mouse genome (mm10) in 1 Kbp bins
}
\keyword{datasets}
