% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enzymes_dm3}
\alias{enzymes_dm3}
\title{Number of restriction enzyme sites for dm3 (flybase)}
\format{A data frame with 130434 rows and 7 variables:
\describe{
  \item{chr}{chromosome}
  \item{position}{start postion of the genomic bin}
  \item{BspHI}{number of RES for BspHI}
  \item{DpnII}{number of RES for DpnII}
  \item{HindIII}{number of RES for HindIII}
  \item{MspI}{number of RES for MspI}
  \item{NcoI}{number of RES for NcoI}
}}
\usage{
enzymes_dm3
}
\description{
A \code{data.frame} containing the number of
restriction enzyme sites (RES) of the fly genome
(dm3, flybase) in 1 Kbp bins. There is one column per
restriction enzyme
}
\details{
Please note that MboI presents the same motif
as DpnII, so use that column corerspondingly
}
\keyword{datasets}
