% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat}
\alias{mat}
\title{HiC contact matrix of a T47D sample at 100 Kbp
resolution}
\format{A 30896 x 330896 sparse \code{dgCMatrix} with the
bin ids as dimnames}
\source{
\url{https://doi.org/10.1101/gad.241422.114}
}
\usage{
mat
}
\description{
A sparse \code{dgCMatrix} containing the HiC contacts
of a T47D sample (SRR1054341) aligned against hg38 at 100 Kbp
resolution. The restriction enzyme used was HindIII. More info at
\url{hhttps://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM1294038}
}
\keyword{datasets}
