% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_k562}
\alias{mat_k562}
\title{HiC contact matrix for Rao's HIC069 sample at 1 Mbp
resolution}
\format{A 3044 x 3044 sparse \code{dgCMatrix} with the
bin ids as dimnames}
\source{
\url{http://dx.doi.org/10.1016/j.cell.2014.11.021}
}
\usage{
mat_k562
}
\description{
A sparse \code{dgCMatrix} containing the HiC contacts
for Rao's sample HIC069 (SRR1658693) at 1 Mbp
resolution. The restriction enzyme used was MboI (same
motif as DpnII). It is a sample of the cell-line K562.
}
\keyword{datasets}
